+++ example/supabase_extensions_example.dart
@@ -1,4 +1,5 @@
 import 'package:supabase/supabase.dart';
+import 'package:supabase_extensions/src/base.dart';
 import 'package:supabase_extensions/src/query_results.dart';
 import 'package:supabase_extensions/supabase_extensions.dart';
 
@@ -9,22 +10,30 @@ void main() async {
 
   // final userID = supabase.uid;
 
-  var results = await supabase.sql(
-      "select * from user_constraints where user_id = 'dcde0dba-f759-4700-84a2-5534aadaaf54'");
+  var userId = "dcde0dba-f759-4700-84a2-5534aadaaf54";
+  //var results = await supabase.sql("select * from user_constraints where user_id = '$u'");
+  //print(results.rows ?? []);
+
+  var results = await supabase.sql("select * from user_constraints where user_id = '$userId'");
   print(results.rows ?? []);
 
   supabase.auth.onAuthStateChange.listen((authState) {
-    if (authState.event == AuthChangeEvent.signedOut) {
+    if(authState.event == AuthChangeEvent.signedOut) {
       // go to login screen automagically
     }
   });
+
 
   //
   // supabase.from('test').stream(primaryKey: ['id']).listen((event) {
   //   print(event);
   // });
 
-  supabase.on('test', 'INSERT').listen((event) {
+  supabase.on('test', CrudEvent.insert).listen((event) {
+    print(event);
+  });
+
+  supabase.onInsert('test').listen((event) {
     print(event);
   });
 
@@ -53,6 +62,7 @@ void main() async {
 
   await Future.delayed(Duration(seconds: 100));
 
+
   // supabase.deleteFolder(bucketId, folderPath);
   // storage.from(bucketId).remove([folderPath]);
 
@@ -61,7 +71,7 @@ void main() async {
   // List<Map<String, dynamic>> results = await supabase.sql(sqlString);
 
   //
-  // QueryResults data =  await supabase.sql("select last_data from saved_schedules where user_id = '6bce5d00-5365-4d44-a33b-089ee431161b'");
+  QueryResults data =  await supabase.sql("select last_data from saved_schedules where user_id = '6bce5d00-5365-4d44-a33b-089ee431161b'");
   // print(data.rows);
 
   // supabase.from('app_counters').update({'value': 6}, options: FetchOptions(forceResponse: true))
@@ -69,4 +79,6 @@ void main() async {
   // final sqlInsertString = "INSERT INTO app_counters (type, value) VALUES ('kuku', 6)";// WHERE type = 'app_visits'";
   // var results = await supabase.sql(sqlInsertString);
   // print(results);
+
+
 }