+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.6.0
+- added `Supabase.isLogged`
+- updated `Supabase.on` event parameter from `String` into enum `CrudEvent`
+- added ability to listen to `DELETE` events
+
 ## 0.5.1
 - Better error handling + updated `pub.dev` score